experimental powergrid changes

+++ gamemodes/basewars/entities/entities/bw_base_generator/power_grid.lua
@@ -1,12 +1,71 @@
 AddCSLuaFile()
 
 PowerGrid = PowerGrid or Networkable:extend()--Emitter:extend()
+PowerGrids = PowerGrids or {}
+
+PowerGridIDsToEnts = PowerGridIDsToEnts or {}
+
+local clprint = CLIENT and BlankFunc or BlankFunc
+local print = SERVER and BlankFunc or BlankFunc
+
+function PowerGridIDsToEnts.Add(ent, id)
+	if not id then id = ent:GetGridID() end
+
+	local grid = PowerGrids[id]
+	if grid and not grid.AllEntities[ent:EntIndex()] then
+		grid:AddEntity(ent)
+	end
+
+	local arr = PowerGridIDsToEnts[id] or {}
+	PowerGridIDsToEnts[id] = arr
+
+	for k,v in ipairs(arr) do
+		if v == ent then
+			return -- we don't need to re-add an entity we already have
+		end
+	end
+
+	arr[#arr + 1] = ent
+end
+
+local function onlyValid(v)
+	return v:IsValid()
+end
+
+function PowerGridIDsToEnts.Sync(id)
+	local entArray = PowerGridIDsToEnts[id]
+	if not entArray then return end
+
+	local grid = PowerGrids[id]
+	if not grid then return end
+
+	table.Filter(entArray, onlyValid)
+
+	for k,v in ipairs(entArray) do
+		local eID = v:EntIndex()
+		if not grid.AllEntities[eID] then
+			grid:AddEntity(v)
+		end
+	end
+
+end
 
 --[[
 
- networked:
-	0: PowerStored
+	networked:
+		0: PowerStored
 
+	emitters:
+		Added[Generator/Line/Consumer]:
+			1 - self: grid
+			2 - entity which got added
+
+		Removed[Generator/Line/Consumer]:
+			1 - self: grid
+			2 - entity which got added
+			3 - was it removed to create a new powergrid for itself?
+
+		GridChanged: when anything in the grid changes (added/removed ent)
 ]]
 function PowerGrid.UpdateIDs()
 	for k, grid in pairs(PowerGrids) do
@@ -16,13 +75,12 @@ function PowerGrid.UpdateIDs()
 
 		for id, ent in pairs(grid.AllEntities) do
 			if not IsValid(ent) then grid.AllEntities[id] = nil continue end --wat
-			if SERVER then 	ent:SetGridID(k) else
-							ent.OldGridID = k end
+			if SERVER then 	ent:SetGridID(k) --[[else
+							ent.OldGridID = k]] end
 		end
 	end
 end
 
-PowerGrids = PowerGrids or {}
 
 function PowerGrid:UpdatePowerIn(gen)
 	local pw_in = 0
@@ -80,6 +138,8 @@ function PowerGrid:Initialize(ow, id, id2) --`id` is only used clientside, when 
 	self.ID = newid
 
 	self:SetNetworkableID("PowerGrid" .. newid)
+
+	__GridEvents = self.__Events
 end
 
 if CLIENT then
@@ -89,8 +149,9 @@ if CLIENT then
 end
 
 PowerGrid:On("GridChanged", "TrackConnections", function(self)
-
-	if self.Connections == 0 then
+	clprint("Grid #" .. self.ID .. " changed; checking...", self.Connections, "connections")
+	if self.Connections <= 0 then
+		clprint("!!! Removed PowerGrid", self.ID, "!!!")
 		self:Remove()
 	end
 
@@ -106,35 +167,71 @@ PowerGrid:On("GridChanged", "UpdateNetworking", function(self, ent)
 end)
 
 PowerGrid:On("RemovedLine", "RepickLine", function(self, line)
+	print("line", line, "removed from grid #" .. self.ID, "; repicking")
+	if CLIENT then return end
+
+	local entsToClosestPoles = {}
+	local allEnts = {}
 
 	for k,v in pairs(self.Accessors) do
-		--if k == "Line" then continue end
+		if k == "Line" then continue end
+
 		local ents = self[v.tbl]
+		for k, ent in ipairs(ents) do
+			local new = self.FindNearestPole(ent, line)
+			entsToClosestPoles[ent] = new
+			allEnts[#allEnts + 1] = ent
+		end
+	end
 
-		for i=#ents, 1, -1 do --ty CornerPin for this https://discordapp.com/channels/565105920414318602/567617926991970306/720349414408847370
-			--if there's no pole to be found here then the entity will be removed from the new grid, and if there's 0 connections
-			--the powergrid will be removed entirely (replaced by new ones), see :On("GridChanged") above for the removal
+	-- reconnect power lines first and collect possible entities to them
+	for k,v in ipairs(self.PowerLines) do
 
-			local ent = ents[i]
-			if ent:GetLine() ~= line then continue end
+		print("what do we do with pole", ent, "?")
+		local new = self.FindNearestPole(v, line)
 
-			local new = self.FindNearestPole(ent)
-
-			if not new then
-				self["Remove" .. k] (self, ent, true) --failed to find a new pole for the entity, disconnect them into their own empty grid
-			else
-				ent.Line = new
-				ent:SetLine(new)
+		if not new then
+			print("didn't find replacement; fuck it")
+			self:RemoveLine(v, true) 
+			-- failed to find a new pole for the entity, disconnect them into their own empty grid
+		else
+			-- found a replacement pole to connect to from the same grid
+			print("found replacement")
+			if new:GetLine() ~= v then -- no circular connections >:(
+				v.Line = new
+				v:SetLine(new)
 			end
-
 		end
 
+		-- then try to grab as many ents to it as possible if the closest pole is that pole
+		local grid = v:GetGrid()
+
+		for ent, pole in pairs(entsToClosestPoles) do
+			if v == pole then
+				if not grid.AllEntities[ent:EntIndex()] then
+					v:GetGrid():AddEntity(ent)
+				end
+
+				ent:SetLine(v)
+				ent.Line = v
+			end
+		end
+	end
+
+	-- the rest (uncollected entities) didn't get a new pole to connect to, therefore, they should be disconnected into their own grid
+
+	for k,v in ipairs(allEnts) do
+		if v:GetLine() == line then
+			v:GetGrid()["Remove" .. v.PowerType] (v:GetGrid(), v, true)
+		end
 	end
 
 end)
 
 function PowerGrid:Remove()
 	PowerGrids[self.ID] = nil
+	PowerGridIDsToEnts[self.ID] = {}
+	clprint("Removing grid", self.ID, " and id->grid too", debug.traceback())
 	self:Invalidate()
 end
 
@@ -194,11 +291,15 @@ for k,v in pairs(accessors) do
 		self:Emit("Added" .. v.emit, ent, ...)
 
 		if SERVER then ent:SetGridID(self.ID) end
+
+		if CLIENT then
+			PowerGridIDsToEnts.Sync(id)
+		end
 	end
 
 	PowerGrid["Remove" .. k] = function(self, ent, new)
 		local t = self[v.tbl]
-		if not self.AllEntities[ent:EntIndex()] then print("Can't remove", ent, "since it wasn't in AllEntities", ent:EntIndex()) return end
+		if not self.AllEntities[ent:EntIndex()] then print("Can't remove", ent, Realm(true, true), "since it wasn't in AllEntities", ent:EntIndex()) return end
 
 		for k,v in ipairs(t) do --remove from PowerType table (generator / powerline / battery)
 			if v == ent then
@@ -209,28 +310,44 @@ for k,v in pairs(accessors) do
 
 		self.AllEntities[ent:EntIndex()] = nil
 
-		--[[for k,v in pairs(self.AllEntities) do --remove from all entities table
+		--[[
+		for k,v in pairs(self.AllEntities) do --remove from all entities table
 			self.AllEntities[ent:EntIndex()] = nil
-		end]]
+		end
 
 		if ent.SetLine then ent:SetLine(NULL) end
+		]]
 
 		self.Connections = self.Connections - 1
 
-		self:Emit("GridChanged", ent)
-		self:Emit("Removed" .. v.emit, ent, new)
+		if CLIENT then 
+			for k,v in ipairs(PowerGridIDsToEnts[self.ID]) do
+				if v == ent then
+					table.remove(PowerGridIDsToEnts[self.ID], k)
+					break
+				end
+			end
 
-		if new then
-			local new = PowerGrid:new(ent:CPPIGetOwner())
-			new["Add" .. ent.PowerType] (new, ent)
-			if SERVER then ent:SetGridID(ent:GetGrid().ID) end
+			--clprint("Removing", ent, "grid #" .. self.ID, self.Connections)
+			--PrintTable(self.AllEntities)
 		end
 
+		self:Emit("GridChanged", ent)
+		self:Emit("Removed" .. v.emit, ent, new)
+		
+
+		if new then
+			PowerGrid:new(ent:CPPIGetOwner()):AddEntity(ent)
+		end
+
+		if CLIENT then
+			PowerGridIDsToEnts.Sync(self.ID)
+		end
 	end
 
 end
 
-function PowerGrid.FindNearestPole(ent) --this isn't a class function, it's a utility function
+function PowerGrid.FindNearestPole(ent, from) --this isn't a class function, it's a utility function
 	local mypos = ent:GetPos()
 	local cable = ent.ConnectDistance ^ 2
 	local ow = ent:CPPIGetOwner()
@@ -239,9 +356,9 @@ function PowerGrid.FindNearestPole(ent) --this isn't a class function, it's a ut
 		grid.Owner = grid.Owner or table.Random(grid.AllEntities):CPPIGetOwner()
 		if not grid.Owner or not grid.Owner:IsValid() or not grid.Owner:IsTeammate(ow) then print("no", grid.Owner) continue end
 		local mindist, minpole = math.huge
-
+		print("probing grid", grid.ID, "for powerlines")
 		for _, line in ipairs(grid.PowerLines) do
-			if line == ent then continue end
+			if line == ent or line == from or not line:IsValid() then print("not considering", line, line == from) continue end
 
 			local pos = line:GetPos()
 			local dist = pos:DistToSqr(mypos)
@@ -249,6 +366,8 @@ function PowerGrid.FindNearestPole(ent) --this isn't a class function, it's a ut
 			if dist < cable and dist < mindist then
 				mindist = dist
 				minpole = line
+			else
+				print("too far", dist, mindist, cable)
 			end
 		end
 
@@ -318,7 +437,7 @@ function PowerGrid:Read()
 	local added = net.ReadUInt(8)
 	for i=1, added do
 		local ent = net.ReadEntity()
-		if not IsValid(ent) then continue end --um?
+		if not ent:IsValid() then print("Invalid entity for grid;", ent) continue end --um?
 
 		self.AllEntities[ent:EntIndex()] = ent
 
@@ -331,6 +450,7 @@ end
 
 hook.Add("EntityRemoved", "ClearGrid", function(ent)
 	local grid = ent.Grid
+	clprint("!!! entity removed:", ent, "!!!")
 	if ent.PowerType and grid then
 		local ok, err = pcall(grid["Remove" .. ent.PowerType], grid, ent)
 		if not ok then
@@ -412,7 +532,7 @@ function PowerGrid:Think()
 			v.RebootStart = CurTime()
 			should_reboot = true
 			v.ShouldReboot = false
-		elseif not enough and not v.Power and not is_rebooting then
+		elseif not enough and not v.Power and not is_rebooting and v.EverPowered then
 			v.ShouldReboot = true --next time it's powered it will go into reboot
 			should_reboot = false
 		end
@@ -429,8 +549,8 @@ function PowerGrid:Think()
 
 
 		--it should resurrect this tick, check reboot time
-		if enough and is_rebooting then
-			-- it's still rebooting; drain power but don't make it powered yet
+		if is_rebooting then
+			-- it's still rebooting; don't make it powered yet
 			enough = false
 		end
 
@@ -503,7 +623,7 @@ else
 		end
 	end)
 
-
 end
+
